merge 'inventory-lootbag-rework' - gd-lootbag v0.8
This is version 0.8 of the Lootbag module.

- Reworked Inventory class.
- Reworked Bag class, renamed to LootBag.
- Added LootItem class: represents a loot bag item.

Renamed some classes to make them less generic, for example Bag versus
LootBag.

Introduced LootItem class, a representation of an item inside the
loot bag. This change is an abstraction for in-game items and
items contained inside the inventory.

Removed some bloat such as useless methods in the Inventory and LootBag
classes.

The initialization process and the structure of the LootBag class is
less complex.

Inventory class is more or less the same, simply adapted the changes
from LootBag/LootItem.

TODO and WIP:

- LootItem is currently just an empty class. v0.9 should be a correct
  implementation of the class. Need to figure it out.

- v1.0 should be the first full implementation of the module. It will include
  how equipment and equipment slots should be handled.